Scene loader, upload buffers, fences, frames and queues

+++ autograph/src/gfx/context.rs
@@ -22,7 +22,7 @@ extern "system" fn debug_callback(
 pub struct ContextConfig
 {
     pub max_frames_in_flight: u32,
-    pub default_upload_buffer_size: u32
+    pub default_upload_buffer_size: usize
 }
 
 #[derive(Debug)]
